% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_profiles.R
\name{aggregate_profiles}
\alias{aggregate_profiles}
\title{aggregate_profiles}
\usage{
aggregate_profiles(new_profiles_table, filter_profile_colnames, supplementary_info_variables, date_colnames, hours_colnames, layer_from_colnames, density_final_colnames, SWE_colnames)
}
\arguments{
\item{new_profiles_table}{Table of snow profiles having Stratigraphy and Density}

\item{filter_profile_colnames}{Name of column of new_profiles_table to use to aggregate profiles (Create before new_profiles_table with other functions in this package)}

\item{supplementary_info_variables}{A vector containing the names of columns to return in output table (Exclude Snow Height, Snow Density and Snow Water Equivalent)}

\item{date_colnames}{Name of column associated to Dates}

\item{hours_colnames}{Name of column associated to Hours}

\item{layer_from_colnames}{Name of column associated to Max Snow Height Layer}

\item{density_final_colnames}{Name of column associated to layer snow density}

\item{SWE_colnames}{Name of column associated to layer Snow Water Equivalent}
}
\value{
A dataframe having that summarise the max snow height, the meand snow density and the total snow water equivalent for each surveys executed
}
\description{
aggregate_profiles() is a function that aggregate by date and station snow profiles obtaining a information of Snow Height total, mean snow density and Snow Water equivalent for each profile.
We suggest to aggregate profiles after a preprocessing using other functions in SnowProfilesProcessing package
}
