% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_density_matrix_grain_hardness.R
\name{calculate_density_matrix_grain_hardness}
\alias{calculate_density_matrix_grain_hardness}
\title{calculate_density_matrix_grain_hardness}
\usage{
calculate_density_matrix_grain_hardness (merged_table, grain_shape1_colnames, grain_shape2_colnames, density_colnames, hardness_colnames)
}
\arguments{
\item{merged_table}{A table containing stratigraphy and density information for each layer}

\item{grain_shape1_colnames}{The name of column associated to grain shape 1}

\item{density_colnames}{The name of column associated to Density}

\item{hardness_colnames}{The name of column associated to Hardness}

\item{grain_shape1_colnames}{The name of column associated to grain shape 2}
}
\value{
A list of 3 elements. The first element is a dataframe of estimated density for each pairing of grain shape and density,
The second element is a dataframe of number of observation for each pairing of grain shape and density
The third element is a dataframe of standard deviation of density for each pairing of grain shape and density
}
\description{
calculate_density_matrix_grain_hardness() is a function that combine grain shapes and hardness to estimate density. The method following paper Valt e Cagnati, "Stima della densita della neve conoscendo la forma dei grani e la durezza", 2005 Neve e Valanghe,55 pag. 40-45
}
