% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator_density_and_SWE.R
\name{calculator_density_and_SWE}
\alias{calculator_density_and_SWE}
\title{calculator_density_and_SWE}
\usage{
calculator_density_and_SWE(merged_table, SWE_grain_table,
  grain_shape1_colnames, grain_shape2_colnames, density_colnames,
  hardness_colnames, layer_from_colnames, layer_to_colnames,
  filter_profile_colnames)
}
\arguments{
\item{merged_table}{A table containing stratigraphy and density information for each layer}

\item{SWE_grain_table}{A table containing snow density for each pairing of grain shape and hardness}

\item{grain_shape1_colnames}{The name of column of merged_table associated to grain shape 1}

\item{density_colnames}{The name of column of merged_table associated to Density}

\item{hardness_colnames}{The name of column of merged_table associated to Hardness}

\item{layer_from_colnames}{The name of column of merged_table associated to the top of layer}

\item{layer_to_colnames}{The name of column of merged_table associated to the bottom of layer}

\item{filter_profile_colnames}{The name of column of merged_tableassociated to filtration code (Usually datetime + station code. We suggest to create merged_table with other funciton in this package)}

\item{grain_shape1_colnames}{The name of column of merged_table associated to grain shape 2}
}
\value{
A list of 3 elements. The first element is a dataframe of estimated density for each pairing of grain shape and density,
The second element is a dataframe of number of observation for each pairing of grain shape and density
The third element is a dataframe of standard deviation of density for each pairing of grain shape and density
}
\description{
calculator_density_and_SWE() is a function that reconstruct snow density for each layer if it is missing using an external matrix of snow density for each pairing of grain shape and hardness (Valt et al.,2005)
}
