% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_stratigraphy_density.R
\name{combine_stratigraphy_density}
\alias{combine_stratigraphy_density}
\title{combine_stratigraphy_density}
\usage{
combine_stratigraphy_density(input_dir, stratigraphy_file, density_file, dates_colnames, hours_colnames, station_code)
}
\arguments{
\item{input_dir}{Folder where stratigraphy_file and density file are stored}

\item{stratigraphy_file}{A file .csv containing stratigraphy for each layer and each profile}

\item{density_file}{A file .csv containing density for each layer and each profile}

\item{dates_colnames}{The name of column associated to Dates}

\item{hours_colnames}{The name of column associated to Hours}

\item{station_code}{The name of column associated to Station identifier}
}
\value{
A dataframe that is the combiantion of stratigraphy file and the density sample extract in density_file
}
\description{
combine_stratigraphy_density() is a function that combine stratigraphy table and density table. The script joins the two table by datetime and station, centering density inside the proper layer
}
