% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_config_xml_parsing.R
\name{file_config_xml_parsing}
\alias{file_config_xml_parsing}
\title{file_config_xml_parsing}
\usage{
file_config_xml_parsing(inpt_file)
}
\arguments{
\item{inpt_file}{a list, the file readed with function xmlParse()}
}
\value{
a list of variables assignemet
}
\description{
This function read file_config.xml file and organize inforamtions in a list
}
\examples{
See an example in inst/examples_profile_analysis.R
}
